split the huge file into its natural modules (#460)
* split the huge file into its natural modules

* remove unused imports
