+++ jenkins/helper/test_launch_controller.py
@@ -1,1086 +1,20 @@
 #!/bin/env python3
 """ read test definition, and generate the output for the specified target """
 import argparse
-import copy
-from datetime import datetime, timedelta
-import platform
-import os
-from pathlib import Path
-import pprint
-import signal
 import sys
-from threading  import Thread, Lock
-import time
 from traceback import print_exc
-import shutil
-import psutil
 
-from async_client import (
-    ArangoCLIprogressiveTimeoutExecutor,
+from site_config import IS_ARM, IS_WINDOWS, IS_MAC
 
-    make_logfile_params,
-    logfile_line_result,
-    delete_logfile_params,
-
-    make_tail_params,
-    tail_line_result,
-    delete_tail_params
-)
-
-ZIPFORMAT="gztar"
-try:
-    import py7zr
-    shutil.register_archive_format('7zip', py7zr.pack_7zarchive, description='7zip archive')
-    ZIPFORMAT="7zip"
-except ModuleNotFoundError:
-    pass
+from dump_handler import generate_dump_output
+from launch_handler import launch
 
 # check python 3
 if sys.version_info[0] != 3:
     print("found unsupported python version ", sys.version_info)
     sys.exit()
 
-
-IS_ARM = platform.processor() == "arm" or platform.processor() == "aarch64"
-IS_WINDOWS = platform.win32_ver()[0] != ""
-IS_MAC = platform.mac_ver()[0] != ""
-IS_LINUX = not IS_MAC and not IS_WINDOWS
-if IS_MAC:
-    # Put us to the performance cores:
-    # https://apple.stackexchange.com/questions/443713
-    from os import setpriority
-    PRIO_DARWIN_THREAD  = 0b0011
-    PRIO_DARWIN_PROCESS = 0b0100
-    PRIO_DARWIN_BG      = 0x1000
-    setpriority(PRIO_DARWIN_PROCESS, 0, 0)
-
-pp = pprint.PrettyPrinter(indent=4)
-
-all_tests = []
 #pylint: disable=line-too-long disable=broad-except disable=chained-comparison
-
-def sigint_boomerang_handler(signum, frame):
-    """do the right thing to behave like linux does"""
-    # pylint: disable=unused-argument
-    if signum != signal.SIGINT:
-        sys.exit(1)
-    # pylint: disable=unnecessary-pass
-    pass
-
-if IS_WINDOWS:
-    original_sigint_handler = signal.getsignal(signal.SIGINT)
-    signal.signal(signal.SIGINT, sigint_boomerang_handler)
-    # pylint: disable=unused-import
-    # this will patch psutil for us:
-    import monkeypatch_psutil
-
-def get_workspace():
-    """ evaluates the directory to put reports to """
-    if 'INNERWORKDIR' in os.environ:
-        workdir = Path(os.environ['INNERWORKDIR'])
-        if workdir.exists():
-            return workdir
-    if 'WORKDIR' in os.environ:
-        workdir = Path(os.environ['WORKDIR'])
-        if workdir.exists():
-            return workdir
-    #if 'WORKSPACE' in os.environ:
-    #    workdir = Path(os.environ['WORKSPACE'])
-    #    if workdir.exists():
-    #        return workdir
-    return Path.cwd() / 'work'
-
-for env in os.environ:
-    print(f'{env}={os.environ[env]}')
-TEMP = Path("/tmp/")
-if 'TMP' in os.environ:
-    TEMP = Path(os.environ['TMP'])
-if 'TEMP' in os.environ:
-    TEMP = Path(os.environ['TEMP'])
-if 'TMP' in os.environ:
-    TEMP = Path(os.environ['TMP'])
-if 'INNERWORKDIR' in os.environ:
-    TEMP = Path(os.environ['INNERWORKDIR'])
-    wd = TEMP / 'ArangoDB'
-    wd.cwd()
-    TEMP = TEMP / 'tmp'
-else:
-    TEMP = TEMP / 'ArangoDB'
-if not TEMP.exists():
-    TEMP.mkdir(parents=True)
-os.environ['TMPDIR'] = str(TEMP)
-os.environ['TEMP'] = str(TEMP)
-os.environ['TMP'] = str(TEMP)
-
-MAX_COREFILES_SINGLE=4
-MAX_COREFILES_CLUSTER=15
-if 'MAX_CORECOUNT' in os.environ:
-    MAX_COREFILES_SINGLE=int(os.environ['MAX_CORECOUNT'])
-    MAX_COREFILES_CLUSTER=int(os.environ['MAX_CORECOUNT'])
-MAX_COREFILE_SIZE_MB=750
-if 'MAX_CORESIZE' in os.environ:
-    MAX_COREFILE_SIZE_MB=int(os.environ['MAX_CORESIZE'])
-
-def get_file_size(elem):
-    """ get the size of a file for sorting """
-    return elem.stat().st_size
-
-def list_all_processes():
-    """list all processes for later reference"""
-    pseaf = "PID  Process"
-    # pylint: disable=catching-non-exception
-    for process in psutil.process_iter(["pid", "name"]):
-        cmdline = process.name
-        try:
-            cmdline = str(process.cmdline())
-            if cmdline == "[]":
-                cmdline = "[" + process.name() + "]"
-        except psutil.AccessDenied:
-            pass
-        except psutil.ProcessLookupError:
-            pass
-        except psutil.NoSuchProcess:
-            pass
-        print(f"{process.pid} {cmdline}")
-    print(pseaf)
-    sys.stdout.flush()
-
-class ArangoshExecutor(ArangoCLIprogressiveTimeoutExecutor):
-    """configuration"""
-
-    def __init__(self, site_config, slot_lock):
-        self.slot_lock = slot_lock
-        self.read_only = False
-        super().__init__(site_config, None)
-
-    def run_testing(self,
-                    testcase,
-                    testing_args,
-                    timeout,
-                    directory,
-                    logfile,
-                    identifier,
-                    verbose
-                    ):
-       # pylint: disable=R0913 disable=R0902
-        """ testing.js wrapper """
-        print('------')
-        print(testing_args)
-        args = [
-            '-c', str(self.cfg.cfgdir / 'arangosh.conf'),
-            "--log.foreground-tty", "true",
-            "--log.force-direct", "true",
-            '--log.level', 'warning',
-            "--log.level", "v8=debug",
-            '--server.endpoint', 'none',
-            '--javascript.allow-external-process-control', 'true',
-            '--javascript.execute', self.cfg.base_path / 'UnitTests' / 'unittest.js',
-            ]
-        run_cmd = args +[
-            '--',
-            testcase,
-            '--testOutput', directory ] + testing_args
-        params = make_logfile_params(verbose, logfile, self.cfg.trace)
-        ret = self.run_monitored(
-            self.cfg.bin_dir / "arangosh",
-            run_cmd,
-            params=params,
-            progressive_timeout=timeout,
-            deadline=self.cfg.deadline,
-            result_line_handler=logfile_line_result,
-            identifier=identifier
-        )
-        delete_logfile_params(params)
-        ret['error'] = params['error']
-        return ret
-
-class DmesgWatcher(ArangoCLIprogressiveTimeoutExecutor):
-    """configuration"""
-
-    def __init__(self, site_config):
-        self.params = None
-        super().__init__(site_config, None)
-
-    def launch(self):
-       # pylint: disable=R0913 disable=R0902
-        """ dmesg wrapper """
-        print('------')
-        args = ['-wT']
-        verbose = False
-        self.params = make_tail_params(verbose, 'dmesg ', self.cfg.test_report_dir / 'dmesg_log.txt')
-        ret = self.run_monitored(
-            "dmesg",
-            args,
-            params=self.params,
-            progressive_timeout=9999999,
-            deadline=self.cfg.deadline,
-            result_line_handler=tail_line_result,
-            identifier='0_dmesg'
-        )
-        #delete_logfile_params(params)
-        ret = {}
-        ret['error'] = self.params['error']
-        delete_tail_params(self.params)
-        return ret
-
-    def end_run(self):
-        """ terminate dmesg again """
-        print(f'killing dmesg {self.pid}')
-        try:
-            psutil.Process(self.pid).kill()
-        except psutil.NoSuchProcess:
-            print('dmesg already gone?')
-            pass
-
-TEST_LOG_FILES = []
-
-class TestConfig():
-    """ setup of one test """
-    # pylint: disable=too-many-instance-attributes disable=too-many-arguments
-    # pylint: disable=too-many-branches disable=too-many-statements
-    # pylint: disable=too-few-public-methods
-    def __init__(self,
-                 cfg,
-                 name,
-                 suite,
-                 args,
-                 priority,
-                 parallelity,
-                 flags):
-        """ defaults for test config """
-        self.parallelity = parallelity
-        self.launch_delay = 1.3
-        self.progressive_timeout = 100
-        self.priority = priority
-        self.suite = suite
-        self.name = name
-        self.name_enum = name
-        self.crashed = False
-        self.success = True
-        self.structured_results = ""
-        self.summary = ""
-        self.start = None
-        self.finish = None
-        self.delta_seconds = 0
-        self.delta = None
-
-        self.base_logdir = cfg.test_report_dir / self.name
-        if not self.base_logdir.exists():
-            self.base_logdir.mkdir()
-        self.log_file =  cfg.run_root / f'{self.name}.log'
-        # pylint: disable=global-variable-not-assigned
-        global TEST_LOG_FILES
-        try:
-            print(TEST_LOG_FILES.index(str(self.log_file)))
-            raise Exception(f'duplicate testfile {str(self.log_file)}')
-        except ValueError:
-            TEST_LOG_FILES.append(str(self.log_file))
-        self.summary_file = self.base_logdir / 'testfailures.txt'
-        self.crashed_file = self.base_logdir / 'UNITTEST_RESULT_CRASHED.json'
-        self.success_file = self.base_logdir / 'UNITTEST_RESULT_EXECUTIVE_SUMMARY.json'
-        self.report_file =  self.base_logdir / 'UNITTEST_RESULT.json'
-        self.base_testdir = cfg.test_data_dir_x / self.name
-
-        self.args = copy.deepcopy(cfg.extra_args)
-        for param in args:
-            if param.startswith('$'):
-                paramname = param[1:]
-                if paramname in os.environ:
-                    self.args += os.environ[paramname].split(' ')
-                else:
-                    print("Error: failed to expand environment variable: '" + param + "' for '" + self.name + "'")
-            else:
-                self.args.append(param)
-        self.args += ['--coreCheck', 'true', '--disableMonitor', 'true', '--writeXmlReport', 'true']
-
-
-        if 'filter' in os.environ:
-            self.args += ['--test', os.environ['filter']]
-        if 'sniff' in flags:
-            if IS_WINDOWS and 'TSHARK' in os.environ:
-                self.args += ['--sniff', 'true',
-                             '--sniffProgram',  os.environ['TSHARK'],
-                             '--sniffDevice', os.environ['DUMPDEVICE']]
-            elif IS_MAC:
-                self.args += ['--sniff', 'sudo']
-            else:
-                self.args += ['--sniff', 'true']
-
-        if 'SKIPNONDETERMINISTIC' in os.environ:
-            self.args += ['--skipNondeterministic', os.environ['SKIPNONDETERMINISTIC']]
-        if 'SKIPTIMECRITICAL' in os.environ:
-            self.args += ['--skipTimeCritical', os.environ['SKIPTIMECRITICAL']]
-
-        if 'BUILDMODE' in os.environ:
-            self.args += [ '--buildType',  os.environ['BUILDMODE'] ]
-
-        if 'DUMPAGENCYONERROR' in os.environ:
-            self.args += [ '--dumpAgencyOnError', os.environ['DUMPAGENCYONERROR']]
-
-        myport = cfg.portbase
-        cfg.portbase += cfg.port_offset
-        self.args += [ '--minPort', str(myport), '--maxPort', str(myport + cfg.port_offset - 1)]
-        if 'SKIPGREY' in os.environ:
-            self.args += [ '--skipGrey', os.environ['SKIPGREY']]
-        if 'ONLYGREY' in os.environ:
-            self.args += [ '--onlyGrey', os.environ['ONLYGREY']]
-
-        if 'vst' in flags:
-            self.args += [ '--vst', 'true']
-        if 'ssl' in flags:
-            self.args += [ '--protocol', 'ssl']
-        if 'http2' in flags:
-            self.args += [ '--http2', 'true']
-
-    def __repr__(self):
-        return f"""
-{self.name} => {self.parallelity}, {self.priority}, {self.success} -- {' '.join(self.args)}"""
-
-    def print_test_log_line(self):
-        """ get visible representation """
-        # pylint: disable=consider-using-f-string
-        resultstr = "Good result in"
-        if not self.success:
-            resultstr = "Bad result in"
-        if self.crashed:
-            resultstr = "Crash occured in"
-        return """
-{1} {0.name} => {0.parallelity}, {0.priority}, {0.success} -- {2}""".format(
-            self,
-            resultstr,
-            ' '.join(self.args))
-
-    def print_testruns_line(self):
-        """ get visible representation """
-        # pylint: disable=consider-using-f-string
-        resultstr = "GOOD"
-        if not self.success:
-            resultstr = "BAD"
-        if self.crashed:
-            resultstr = "CRASH"
-        return """
-<tr><td>{0.name}</td><td align="right">{0.delta}</td><td align="right">{1}</td></tr>""".format(
-            self,
-            resultstr)
-
-def get_priority(test_config):
-    """ sorter function to return the priority """
-    return test_config.priority
-
-class SiteConfig:
-    """ this environment - adapted to oskar defaults """
-    # pylint: disable=too-few-public-methods disable=too-many-instance-attributes
-    def __init__(self, definition_file):
-        # pylint: disable=too-many-statements disable=too-many-branches
-        self.datetime_format = "%Y-%m-%dT%H%M%SZ"
-        self.trace = False
-        self.portbase = 7000
-        if 'PORTBASE' in os.environ:
-            self.portbase = int(os.environ['PORTBASE'])
-        self.port_offset = 100
-        self.timeout = 1800
-        if 'timeLimit'.upper() in os.environ:
-            self.timeout = int(os.environ['timeLimit'.upper()])
-        elif 'timeLimit' in os.environ:
-            self.timeout = int(os.environ['timeLimit'])
-        self.small_machine = False
-        self.extra_args = []
-        if psutil.cpu_count(logical=False) <= 12:
-            print("Small machine detected, quadrupling deadline, disabling buckets!")
-            self.small_machine = True
-            self.port_offset = 400
-            self.timeout *= 4
-        self.no_threads = psutil.cpu_count()
-        self.available_slots = round(self.no_threads * 2) #logical=False)
-        if IS_MAC and platform.processor() == "arm":
-            if psutil.cpu_count() == 8:
-                self.no_threads = 6 # M1 mac mini only has 4 performance cores
-                self.available_slots = 10
-            if psutil.cpu_count() == 20:
-                self.no_threads = 16 # M2 mac studio only has 16 performance cores
-                self.available_slots = 14
-                self.timeout *= 2
-        if IS_WINDOWS:
-            self.max_load = self.no_threads * 0.5
-            self.max_load1 = self.no_threads * 0.6
-        else:
-            self.max_load = self.no_threads * 0.9
-            self.max_load1 = self.no_threads * 1.1
-        # roughly increase 1 per ten cores
-        self.core_dozend = round(self.no_threads / 10)
-        if self.core_dozend == 0:
-            self.core_dozend = 1
-        self.loop_sleep = round(5 / self.core_dozend)
-        self.overload = self.max_load * 1.4
-        self.slots_to_parallelity_factor = self.max_load / self.available_slots
-        self.rapid_fire = round(self.available_slots / 10)
-        self.is_asan = 'SAN' in os.environ and os.environ['SAN'] == 'On'
-        if self.is_asan:
-            print('SAN enabled, reducing possible system capacity')
-            self.rapid_fire = 1
-            self.available_slots /= 4
-            #self.timeout *= 1.5
-            self.loop_sleep *= 2
-            self.max_load /= 2
-            if os.environ['SAN_MODE'] == 'AULSan':
-                print('Aulsan must reduce even more!')
-        self.deadline = datetime.now() + timedelta(seconds=self.timeout)
-        self.hard_deadline = datetime.now() + timedelta(seconds=self.timeout + 660)
-        if definition_file.is_file():
-            definition_file = definition_file.parent
-        base_source_dir = (definition_file / '..').resolve()
-        bin_dir = (base_source_dir / 'build' / 'bin').resolve()
-        if IS_WINDOWS:
-            for target in ['RelWithdebInfo', 'Debug']:
-                if (bin_dir / target).exists():
-                    bin_dir = bin_dir / target
-        socket_count = "was not allowed to see socket counts!"
-        try:
-            socket_count = str(get_socket_count())
-        except psutil.AccessDenied:
-            pass
-
-        print(f"""Machine Info:
- - {psutil.cpu_count(logical=False)} Cores / {psutil.cpu_count(logical=True)} Threads
- - {platform.processor()} processor architecture
- - {psutil.virtual_memory()} virtual Memory
- - {self.slots_to_parallelity_factor} parallelity to load estimate factor
- - {self.overload} load1 threshhold for overload logging
- - {self.max_load} / {self.max_load1} configured maximum load 0 / 1
- - {self.slots_to_parallelity_factor} parallelity to load estimate factor
- - {self.available_slots} test slots {self.rapid_fire} rapid fire slots
- - {str(TEMP)} - temporary directory
- - current Disk I/O: {str(psutil.disk_io_counters())}
- - current Swap: {str(psutil.swap_memory())}
- - Starting {str(datetime.now())} soft deadline will be: {str(self.deadline)} hard deadline will be: {str(self.hard_deadline)}
- - {self.core_dozend} / {self.loop_sleep} machine size / loop frequency
- - {socket_count} number of currently active tcp sockets
- """)
-        self.cfgdir = base_source_dir / 'etc' / 'relative'
-        self.bin_dir = bin_dir
-        self.base_path = base_source_dir
-        self.test_data_dir = base_source_dir
-        self.passvoid = ''
-        self.run_root = base_source_dir / 'testrun'
-        if self.run_root.exists():
-            shutil.rmtree(self.run_root)
-        self.test_data_dir_x = self.run_root / 'run'
-        self.test_data_dir_x.mkdir(parents=True)
-        self.test_report_dir = self.run_root / 'report'
-        self.test_report_dir.mkdir(parents=True)
-        self.portbase = 7000
-        if 'PORTBASE' in os.environ:
-            self.portbase = int(os.environ['PORTBASE'])
-
-    def get_overload(self):
-        """ estimate whether the system is overloaded """
-        load = psutil.getloadavg()
-        if load[0] > self.overload:
-            return(f"HIGH SYSTEM LOAD! {load[0]}")
-
-def dmesg_runner(dmesg):
-    """ thread to run dmesg in """
-    dmesg.launch()
-
-def testing_runner(testing_instance, this, arangosh):
-    """ operate one makedata instance """
-    this.start = datetime.now(tz=None)
-    ret = arangosh.run_testing(this.suite,
-                               this.args,
-                               999999999,
-                               this.base_logdir,
-                               this.log_file,
-                               this.name_enum,
-                               True) #verbose?
-    this.success = (
-        not ret["progressive_timeout"] or
-        not ret["have_deadline"] or
-        ret["rc_exit"] == 0
-    )
-    this.finish = datetime.now(tz=None)
-    this.delta = this.finish - this.start
-    this.delta_seconds = this.delta.total_seconds()
-    print(f'done with {this.name_enum}')
-    this.crashed = not this.crashed_file.exists() or this.crashed_file.read_text() == "true"
-    this.success = this.success and this.success_file.exists() and this.success_file.read_text() == "true"
-    if this.report_file.exists():
-        this.structured_results = this.report_file.read_text(encoding="UTF-8", errors='ignore')
-    this.summary = ret['error']
-    if this.summary_file.exists():
-        this.summary += this.summary_file.read_text()
-    else:
-        print(f'{this.name_enum} no testreport!')
-
-    with arangosh.slot_lock:
-        testing_instance.running_suites.remove(this.name_enum)
-
-    if this.crashed or not this.success:
-        print(str(this.log_file.name))
-        print(this.log_file.parent / ("FAIL_" + str(this.log_file.name))
-              )
-        failname = this.log_file.parent / ("FAIL_" + str(this.log_file.name))
-        this.log_file.rename(failname)
-        this.log_file = failname
-        if (this.summary == "" and failname.stat().st_size < 1024*10):
-            print("pulling undersized test output into testfailures.txt")
-            this.summary = failname.read_text(encoding='utf-8')
-        with arangosh.slot_lock:
-            if this.crashed:
-                testing_instance.crashed = True
-            testing_instance.success = False
-    testing_instance.done_job(this.parallelity)
-
-INTERESTING_SOCKETS = [
-    psutil.CONN_FIN_WAIT1,
-    psutil.CONN_FIN_WAIT2,
-    psutil.CONN_CLOSE_WAIT,
-    psutil.CONN_ESTABLISHED,
-    psutil.CONN_SYN_SENT,
-    psutil.CONN_SYN_RECV,
-    psutil.CONN_TIME_WAIT,
-    psutil.CONN_CLOSE,
-    psutil.CONN_LAST_ACK,
-    psutil.CONN_LISTEN,
-    psutil.CONN_CLOSING
-]
-
-def get_socket_count():
-    """ get the number of sockets lingering destruction """
-    # pylint: disable=too-many-nested-blocks
-    counter = 0
-    if IS_MAC:
-        # Mac would need root for all sockets, so we just look
-        # for arangods and their ports, which works without.
-        for proc in psutil.process_iter(['pid', 'name']):
-            if proc.name() in ['arangod', 'arangosh']:
-                try:
-                    for socket in psutil.Process(proc.pid).connections():
-                        if socket.status in INTERESTING_SOCKETS:
-                            counter += 1
-                except psutil.ZombieProcess:
-                    pass
-                except psutil.NoSuchProcess:
-                    pass
-    else:
-        for socket in psutil.net_connections(kind='inet'):
-            if socket.status in INTERESTING_SOCKETS:
-                counter += 1
-    return counter
-
-class TestingRunner():
-    """ manages test runners, creates report """
-    # pylint: disable=too-many-instance-attributes
-    def __init__(self, cfg):
-        self.cfg = cfg
-        self.deadline_reached = False
-        self.slot_lock = Lock()
-        self.used_slots = 0
-        self.scenarios = []
-        self.arangosh = ArangoshExecutor(self.cfg, self.slot_lock)
-        self.workers = []
-        self.running_suites = []
-        self.success = True
-        self.crashed = False
-        self.cluster = False
-        self.datetime_format = "%Y-%m-%dT%H%M%SZ"
-        self.testfailures_file = get_workspace() / 'testfailures.txt'
-        self.overload_report_file = self.cfg.test_report_dir / 'overload.txt'
-        self.overload_report_fh = self.overload_report_file.open('w', encoding='utf-8')
-
-    def print_active(self):
-        """ output currently active testsuites """
-        now = datetime.now(tz=None).strftime(f"testreport-{self.cfg.datetime_format}")
-        load = psutil.getloadavg()
-        used_slots = ""
-        running_slots = ""
-        with self.slot_lock:
-            used_slots = str(self.used_slots)
-            running_slots = str(self.running_suites)
-        print(str(load) + "<= Load " +
-              "Running: " + str(self.running_suites) +
-              " => Active Slots: " + used_slots +
-              " => Swap: " + str(psutil.swap_memory()) +
-              " => Disk I/O: " + str(psutil.disk_io_counters()))
-        sys.stdout.flush()
-        if load[0] > self.cfg.overload:
-            print(f"{now} {load[0]} | {used_slots} | {running_slots}", file=self.overload_report_fh)
-
-    def done_job(self, parallelity):
-        """ if one job is finished... """
-        with self.slot_lock:
-            self.used_slots -= parallelity
-
-    def launch_next(self, offset, counter, do_loadcheck):
-        """ launch one testing job """
-        if do_loadcheck:
-            if self.scenarios[offset].parallelity > (self.cfg.available_slots - self.used_slots):
-                print("no more slots available")
-                return -1
-            try:
-                sock_count = get_socket_count()
-                if sock_count > 8000:
-                    print(f"Socket count: {sock_count}, waiting before spawning more")
-                    return -1
-            except psutil.AccessDenied:
-                pass
-            load_estimate = self.cfg.slots_to_parallelity_factor * self.scenarios[offset].parallelity
-            load = psutil.getloadavg()
-            if ((load[0] > self.cfg.max_load) or
-                (load[1] > self.cfg.max_load1) or
-                (load[0] + load_estimate > self.cfg.overload)):
-                print(F"{str(load)} <= {load_estimate} Load to high; waiting before spawning more - Disk I/O: " +
-                      str(psutil.swap_memory()))
-                return -1
-        parallelity = 0
-        with self.slot_lock:
-            parallelity = self.scenarios[offset].parallelity
-        self.used_slots += parallelity
-        this = self.scenarios[offset]
-        this.name_enum = f"{this.name} {str(counter)}"
-        print(f"launching {this.name_enum}")
-        pp.pprint(this)
-
-        with self.slot_lock:
-            self.running_suites.append(this.name_enum)
-
-        worker = Thread(target=testing_runner,
-                        args=(self,
-                              this,
-                              self.arangosh))
-        worker.name = this.name
-        worker.start()
-        self.workers.append(worker)
-        return parallelity
-
-    def handle_deadline(self):
-        """ here we make sure no worker thread is stuck during its extraordinary shutdown """
-        # 5 minutes for threads to clean up their stuff, else we consider them blocked
-        more_running = True
-        mica = None
-        print(f"Main: {str(datetime.now())} soft deadline reached: {str(self.cfg.deadline)} now waiting for hard deadline {str(self.cfg.hard_deadline)}")
-        while ((datetime.now() < self.cfg.hard_deadline) and more_running):
-            time.sleep(1)
-            with self.slot_lock:
-                more_running = self.used_slots != 0
-        if more_running:
-            print("Main: reaching hard Time limit!")
-            list_all_processes()
-            mica = os.getpid()
-            myself = psutil.Process(mica)
-            children = myself.children(recursive=True)
-            for one_child in children:
-                if one_child.pid != mica:
-                    try:
-                        print(f"Main: killing {one_child.name()} - {str(one_child.pid)}")
-                        one_child.resume()
-                    except psutil.NoSuchProcess:
-                        pass
-                    except psutil.AccessDenied:
-                        pass
-                    try:
-                        one_child.kill()
-                    except psutil.NoSuchProcess:  # pragma: no cover
-                        pass
-            print("Main: waiting for the children to terminate")
-            psutil.wait_procs(children, timeout=20)
-            print("Main: giving workers 20 more seconds to exit.")
-            time.sleep(60)
-            with self.slot_lock:
-                more_running = self.used_slots != 0
-        else:
-            print("Main: workers terminated on time")
-        if more_running:
-            self.generate_report_txt()
-            print("Main: force-terminates the python process due to overall unresponsiveness! Geronimoooo!")
-            list_all_processes()
-            sys.stdout.flush()
-            self.success = False
-            if IS_WINDOWS:
-                # pylint: disable=protected-access
-                # we want to exit without waiting for threads:
-                os._exit(4)
-            else:
-                os.kill(mica, signal.SIGKILL)
-                sys.exit(4)
-
-    def testing_runner(self):
-        """ run testing suites """
-        # pylint: disable=too-many-branches disable=too-many-statements
-        mem = psutil.virtual_memory()
-        os.environ['ARANGODB_OVERRIDE_DETECTED_TOTAL_MEMORY'] = str(int((mem.total * 0.8) / 9))
-
-        start_offset = 0
-        used_slots = 0
-        counter = 0
-        if len(self.scenarios) == 0:
-            raise Exception("no valid scenarios loaded")
-        some_scenario = self.scenarios[0]
-        if not some_scenario.base_logdir.exists():
-            some_scenario.base_logdir.mkdir()
-        if not some_scenario.base_testdir.exists():
-            some_scenario.base_testdir.mkdir()
-        print(self.cfg.deadline)
-        parallelity = 0
-        sleep_count = 0
-        last_started_count = -1
-        if datetime.now() > self.cfg.deadline:
-            raise ValueError("test already timed out before started?")
-        while (datetime.now() < self.cfg.deadline) and (start_offset < len(self.scenarios) or used_slots > 0):
-            used_slots = 0
-            with self.slot_lock:
-                used_slots = self.used_slots
-            if ((self.cfg.available_slots > used_slots) and
-                (start_offset < len(self.scenarios)) and
-                 ((last_started_count < 0) or
-                  (sleep_count - last_started_count > parallelity)) ):
-                print(f"Launching more: {self.cfg.available_slots} > {used_slots} {counter} {last_started_count} ")
-                sys.stdout.flush()
-                rapid_fire = 0
-                par = 1
-                while (self.cfg.rapid_fire > rapid_fire and
-                       self.cfg.available_slots > used_slots and
-                       len(self.scenarios) > start_offset and
-                       par > 0):
-                    par =  self.launch_next(start_offset, counter, last_started_count != -1)
-                    rapid_fire += par
-                    if par > 0:
-                        counter += 1
-                        time.sleep(0.1)
-                        start_offset += 1
-                if par > 0:
-                    parallelity = par
-                    last_started_count = sleep_count
-                    sleep_count += 1
-                    time.sleep(self.cfg.loop_sleep)
-                    self.print_active()
-                else:
-                    if used_slots == 0 and start_offset >= len(self.scenarios):
-                        print("done")
-                        break
-                    self.print_active()
-                    sleep_count += 1
-                    time.sleep(self.cfg.loop_sleep)
-                    sleep_count += 1
-            else:
-                self.print_active()
-                time.sleep(self.cfg.loop_sleep)
-                sleep_count += 1
-        self.deadline_reached = datetime.now() > self.cfg.deadline
-        if self.deadline_reached:
-            self.handle_deadline()
-        for worker in self.workers:
-            if self.deadline_reached:
-                print("Deadline: Joining threads of " + worker.name)
-            worker.join()
-        if self.success:
-            for scenario in self.scenarios:
-                if not scenario.success:
-                    self.success = False
-
-    def generate_report_txt(self):
-        """ create the summary testfailures.txt from all bits """
-        print(self.scenarios)
-        summary = ""
-        if self.deadline_reached:
-            summary = "Deadline reached during test execution!\n"
-        for testrun in self.scenarios:
-            print(testrun)
-            if testrun.crashed or not testrun.success:
-                summary += f"\n=== {testrun.name} ===\n{testrun.summary}"
-            if testrun.finish is None:
-                summary += f"\n=== {testrun.name} ===\nhasn't been launched at all!"
-        print(summary)
-        self.testfailures_file.write_text(summary)
-
-    def append_report_txt(self, text):
-        """ if the file has already been written, but we have more to say: """
-        with self.testfailures_file.open("a") as filep:
-            filep.write(text + '\n')
-
-    def cleanup_unneeded_binary_files(self):
-        """ delete all files not needed for the crashreport binaries """
-        shutil.rmtree(str(self.cfg.bin_dir / 'tzdata'))
-        needed = [
-            'arangod',
-            'arangosh',
-            'arangodump',
-            'arangorestore',
-            'arangoimport',
-            'arangobackup',
-            'arangodbtests']
-        for one_file in self.cfg.bin_dir.iterdir():
-            if (one_file.suffix == '.lib' or
-                (one_file.stem not in needed) ):
-                print(f'Deleting {str(one_file)}')
-                one_file.unlink(missing_ok=True)
-
-    def generate_crash_report(self):
-        """ crash report zips """
-        # pylint: disable=too-many-statements disable=too-many-branches disable=too-many-locals
-        core_max_count = MAX_COREFILES_SINGLE
-        if self.cluster:
-            core_max_count = MAX_COREFILES_CLUSTER
-        core_dir = Path.cwd()
-        core_pattern = "core*"
-        move_files = False
-        if IS_WINDOWS:
-            core_pattern = "*.dmp"
-        system_corefiles = []
-        if 'COREDIR' in os.environ:
-            core_dir = Path(os.environ['COREDIR'])
-        elif IS_LINUX:
-            core_pattern = Path('/proc/sys/kernel/core_pattern').read_text(encoding="utf-8").strip()
-            if core_pattern.startswith('/'):
-                core_dir = Path(core_pattern).parent
-            move_files = True
-        else:
-            move_files = True
-            core_dir = Path('/var/tmp/') # default to coreDirectory in testing.js
-        if IS_MAC:
-            move_files = True
-            system_corefiles = list(Path('/cores').glob(core_pattern))
-            if system_corefiles is None:
-                system_corefiles = []
-        files_unsorted = list(core_dir.glob(core_pattern))
-        if files_unsorted is None:
-            files_unsorted = []
-        files_unsorted += system_corefiles
-        if len(files_unsorted) == 0 or core_max_count <= 0:
-            print(f'Coredumps are not collected: {str(len(files_unsorted))} coredumps found; coredumps max limit to collect is {str(core_max_count)}!')
-            return
-
-        for one_file in files_unsorted:
-            if one_file.is_file():
-                size = (one_file.stat().st_size / (1024 * 1024))
-                if 0 < MAX_COREFILE_SIZE_MB and MAX_COREFILE_SIZE_MB < size:
-                    print(f'deleting coredump {str(one_file)} its too big: {str(size)}')
-                    files_unsorted.remove(one_file)
-            else:
-                files_unsorted.remove(one_file)
-
-        if len(files_unsorted) > core_max_count and core_max_count > 0:
-            count = 0
-            for one_crash_file in files_unsorted:
-                count += 1
-                if count > core_max_count:
-                    print(f'{core_max_count} reached. will not archive {one_crash_file}')
-                    one_crash_file.unlink(missing_ok=True)
-
-        is_empty = len(files_unsorted) == 0
-        if not is_empty and move_files:
-            core_dir = core_dir / 'coredumps'
-            core_dir.mkdir(parents=True, exist_ok=True)
-            for one_file in files_unsorted:
-                if one_file.exists():
-                    try:
-                        shutil.move(str(one_file.resolve()), str(core_dir.resolve()))
-                    except PermissionError as ex:
-                        print(f"won't move {str(one_file)} - not an owner! {str(ex)}")
-                        self.append_report_txt(f"won't move {str(one_file)} - not an owner! {str(ex)}")
-
-        if self.crashed or not is_empty:
-            crash_report_file = get_workspace() / datetime.now(tz=None).strftime(f"crashreport-{self.cfg.datetime_format}")
-            print("creating crashreport: " + str(crash_report_file))
-            sys.stdout.flush()
-            try:
-                shutil.make_archive(str(crash_report_file),
-                                    ZIPFORMAT,
-                                    (core_dir / '..').resolve(),
-                                    core_dir.name,
-                                    True)
-            except Exception as ex:
-                print("Failed to create binaries zip: " + str(ex))
-                self.append_report_txt("Failed to create binaries zip: " + str(ex))
-            self.cleanup_unneeded_binary_files()
-            binary_report_file = get_workspace() / datetime.now(tz=None).strftime(f"binaries-{self.cfg.datetime_format}")
-            print("creating crashreport binary support zip: " + str(binary_report_file))
-            sys.stdout.flush()
-            try:
-                shutil.make_archive(str(binary_report_file),
-                                    ZIPFORMAT,
-                                    (self.cfg.bin_dir / '..').resolve(),
-                                    self.cfg.bin_dir.name,
-                                    True)
-            except Exception as ex:
-                print("Failed to create crashdump zip: " + str(ex))
-                self.append_report_txt("Failed to create crashdump zip: " + str(ex))
-            for corefile in core_dir.glob(core_pattern):
-                print("Deleting corefile " + str(corefile))
-                sys.stdout.flush()
-                corefile.unlink()
-            if not is_empty and move_files:
-                core_dir.rmdir()
-
-    def generate_test_report(self):
-        """ regular testresults zip """
-        tarfile = get_workspace() / datetime.now(tz=None).strftime(f"testreport-{self.cfg.datetime_format}")
-        print("Creating " + str(tarfile))
-        sys.stdout.flush()
-        try:
-            shutil.make_archive(self.cfg.run_root / 'innerlogs',
-                                ZIPFORMAT,
-                                (TEMP / '..').resolve(),
-                                TEMP.name)
-        except Exception as ex:
-            print("Failed to create inner zip: " + str(ex))
-            self.append_report_txt("Failed to create inner zip: " + str(ex))
-            self.success = False
-
-        try:
-            shutil.rmtree(TEMP, ignore_errors=False)
-            shutil.make_archive(tarfile,
-                                ZIPFORMAT,
-                                self.cfg.run_root,
-                                '.',
-                                True)
-        except Exception as ex:
-            print("Failed to create testreport zip: " + str(ex))
-            self.append_report_txt("Failed to create testreport zip: " + str(ex))
-            self.success = False
-        try:
-            shutil.rmtree(self.cfg.run_root, ignore_errors=False)
-        except Exception as ex:
-            print("Failed to clean up: " + str(ex))
-            self.append_report_txt("Failed to clean up: " + str(ex))
-            self.success = False
-
-    def create_log_file(self):
-        """ create the log file with the stati """
-        logfile = get_workspace() / 'test.log'
-        with open(logfile, "w", encoding="utf-8") as filep:
-            for one_scenario in self.scenarios:
-                filep.write(one_scenario.print_test_log_line())
-
-    def create_testruns_file(self):
-        """ create the log file with the stati """
-        logfile = get_workspace() / 'testRuns.html'
-        state = 'GOOD'
-        if not self.success:
-            state  = 'BAD'
-        if self.crashed:
-            state = 'CRASHED'
-        with open(logfile, "w", encoding="utf-8") as filep:
-            filep.write('''
-<table>
-<tr><th>Test</th><th>Runtime</th><th>Status</th></tr>
-''')
-            total = 0
-            for one_scenario in self.scenarios:
-                filep.write(one_scenario.print_testruns_line())
-                total += one_scenario.delta_seconds
-            filep.write(f'''
-<tr style="background-color: red;color: white;"><td>TOTAL</td><td align="right"></td><td align="right">{state}</td></tr>
-</table>
-''')
-
-    def register_test_func(self, cluster, test):
-        """ print one test function """
-        args = test["args"]
-        params = test["params"]
-        suffix = params.get("suffix", "")
-        name = test["name"]
-        if suffix:
-            name += f"_{suffix}"
-
-        if test["parallelity"] :
-            parallelity = test["parallelity"]
-        if 'single' in test['flags'] and cluster:
-            return
-        if 'cluster' in test['flags'] and not cluster:
-            return
-        if cluster:
-            self.cluster = True
-            if parallelity == 1:
-                parallelity = 4
-            args += ['--cluster', 'true',
-                     '--dumpAgencyOnError', 'true']
-        if "enterprise" in test["flags"]:
-            return
-        if "ldap" in test["flags"] and not 'LDAPHOST' in os.environ:
-            return
-
-        if "buckets" in params and not self.cfg.small_machine:
-            num_buckets = int(params["buckets"])
-            for i in range(num_buckets):
-                self.scenarios.append(
-                    TestConfig(self.cfg,
-                               name + f"_{i}",
-                               test["name"],
-                               [ *args,
-                                 '--index', f"{i}",
-                                 '--testBuckets', f'{num_buckets}/{i}'],
-                               test['priority'],
-                               parallelity,
-                               test['flags']))
-        else:
-            self.scenarios.append(
-                TestConfig(self.cfg,
-                           name,
-                           test["name"],
-                           [ *args],
-                           test['priority'],
-                           parallelity,
-                           test['flags']))
-
-    def sort_by_priority(self):
-        """ sort the tests by their priority for the excecution """
-        self.scenarios.sort(key=get_priority, reverse=True)
-
-    def print_and_exit_closing_stance(self):
-        """ our finaly good buye stance. """
-        print("\n" + "SUCCESS" if self.success else "FAILED")
-        retval = 0
-        if not self.success:
-            retval = 1
-        if self.crashed:
-            retval = 2
-        sys.exit(retval)
-
-def launch(args, tests):
-    """ Manage test execution on our own """
-    runner = None
-    try:
-        runner = TestingRunner(SiteConfig(Path(args.definitions).resolve()))
-    except Exception as exc:
-        print(exc)
-        raise exc
-    dmesg = DmesgWatcher(runner.cfg)
-    if IS_LINUX:
-        dmesg_thread = Thread(target=dmesg_runner, args=[dmesg])
-        dmesg_thread.start()
-        time.sleep(3)
-    for test in tests:
-        runner.register_test_func(args.cluster, test)
-    runner.sort_by_priority()
-    print(runner.scenarios)
-    create_report = True
-    if args.no_report:
-        print("won't generate report as you demanded!")
-        create_report = False
-    try:
-        runner.testing_runner()
-        runner.overload_report_fh.close()
-        runner.generate_report_txt()
-        if create_report:
-            runner.generate_test_report()
-            if not runner.cfg.is_asan:
-                runner.generate_crash_report()
-    except Exception as exc:
-        runner.success = False
-        sys.stderr.flush()
-        sys.stdout.flush()
-        print(exc, file=sys.stderr)
-        print_exc()
-    finally:
-        sys.stderr.flush()
-        sys.stdout.flush()
-        runner.create_log_file()
-        runner.create_testruns_file()
-        if IS_LINUX:
-            dmesg.end_run()
-            print('joining dmesg threads')
-            dmesg_thread.join()
-        runner.print_and_exit_closing_stance()
 
 def filter_tests(args, tests):
     """ filter testcase by operations target Single/Cluster/full """
@@ -1117,23 +51,6 @@ def filter_tests(args, tests):
     for one_filter in filters:
         tests = filter(one_filter, tests)
     return list(tests)
-
-
-def generate_dump_output(_, tests):
-    """ interactive version output to inspect comprehension """
-    def output(line):
-        """ output one line """
-        print(line)
-
-    for test in tests:
-        params = " ".join(f"{key}={value}" for key, value in test['params'].items())
-        output(f"{test['name']}")
-        output(f"\tpriority: {test['priority']}")
-        output(f"\tparallelity: {test['parallelity']}")
-        output(f"\tflags: {' '.join(test['flags'])}")
-        output(f"\tparams: {params}")
-        output(f"\targs: {' '.join(test['args'])}")
-
 
 formats = {
     "dump": generate_dump_output,
